% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartialNetwork.R
\docType{package}
\name{PartialNetwork2-package}
\alias{PartialNetwork2}
\alias{PartialNetwork2-package}
\title{The PartialNetwork2 package}
\description{
The \pkg{PartialNetwork2} package implements a peudo Maximum likelihood for the linear-in-mean peer effect model with group heterogeneity under partial network data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ahoundetoungan/PartialNetwork2}
  \item Report bugs at \url{https://github.com/ahoundetoungan/PartialNetwork2/issues}
}

}
\author{
\strong{Maintainer}: Elysee Aristide Houndetoungan \email{ariel92and@gmail.com}

}
